+++ test/automation/src/extensions.ts
@@ -5,6 +5,10 @@
 
 import { Viewlet } from './viewlet';
 import { Code } from './code';
+import path = require('path');
+import fs = require('fs');
+import { ncp } from 'ncp';
+import { promisify } from 'util';
 
 const SEARCH_BOX = 'div.extensions-viewlet[id="workbench.view.extensions"] .monaco-editor textarea';
 
@@ -49,5 +53,13 @@ export class Extensions extends Viewlet {
 			await this.code.waitForElement(`.extension-editor .monaco-action-bar .action-item:not(.disabled) .extension-action[title="Disable this extension"]`);
 		}
 	}
+}
 
+export async function copyExtension(repoPath: string, extensionsPath: string, extId: string): Promise<void> {
+	const dest = path.join(extensionsPath, extId);
+	if (!fs.existsSync(dest)) {
+		const orig = path.join(repoPath, 'extensions', extId);
+
+		return promisify(ncp)(orig, dest);
+	}
 }
